+++ cli/openbb_cli/argparse_translator/argparse_translator.py
@@ -20,176 +20,17 @@ from typing import (
 )
 
 from openbb_core.app.model.field import OpenBBField
-from pydantic import BaseModel, model_validator
+from pydantic import BaseModel
 from typing_extensions import Annotated
+
+from openbb_cli.argparse_translator.argparse_argument import (
+    ArgparseArgumentGroupModel,
+    ArgparseArgumentModel,
+)
 
 # pylint: disable=protected-access
 
 SEP = "__"
-
-
-class ArgparseArgumentModel(BaseModel):
-    """Pydantic model for an argparse argument."""
-
-    name: str
-    type: Any
-    dest: str
-    default: Any
-    required: bool
-    action: Literal["store_true", "store"]
-    help: Optional[str]
-    nargs: Optional[Literal["+"]]
-    choices: Optional[Tuple]
-
-    @model_validator(mode="after")  # type: ignore
-    @classmethod
-    def validate_action(cls, values: "ArgparseArgumentModel"):
-        """Validate the action based on the type."""
-        if values.type is bool and values.action != "store_true":
-            raise ValueError('If type is bool, action must be "store_true"')
-        return values
-
-    @model_validator(mode="after")  # type: ignore
-    @classmethod
-    def remove_props_on_store_true(cls, values: "ArgparseArgumentModel"):
-        """Remove type, nargs, and choices if action is store_true."""
-        if values.action == "store_true":
-            values.type = None
-            values.nargs = None
-            values.choices = None
-        return values
-
-    # override
-    def model_dump(self, **kwargs):
-        """Override the model_dump method to remove empty choices."""
-        res = super().model_dump(**kwargs)
-
-        # Check if choices is present and if it's an empty tuple remove it
-        if "choices" in res and not res["choices"]:
-            del res["choices"]
-
-        return res
-
-
-class ArgparseArgumentGroupModel(BaseModel):
-    """Pydantic model for a custom argument group."""
-
-    name: str
-    arguments: List[ArgparseArgumentModel]
-
-
-class ReferenceToArgumentsProcessor:
-    """Class to process the reference and build custom argument groups."""
-
-    def __init__(self, reference: Dict[str, Dict]):
-        """Initialize the ReferenceToArgumentsProcessor."""
-        self.reference = reference
-        self.custom_groups: Dict[str, List[ArgparseArgumentGroupModel]] = {}
-
-        self.build_custom_groups()
-
-    @staticmethod
-    def _make_type_parsable(type_: str) -> type:
-        """Make the type parsable by removing the annotations."""
-        if "Union" in type_ and "str" in type_:
-            return str
-        if "Union" in type_ and "int" in type_:
-            return int
-        if type_ in ["date", "datetime.time", "time"]:
-            return str
-
-        if any(x in type_ for x in ["gt=", "ge=", "lt=", "le="]):
-            if "Annotated" in type_:
-                type_ = type_.replace("Annotated[", "").replace("]", "")
-            type_ = type_.split(",")[0]
-
-        return eval(type_)  # noqa: S307, E501 pylint: disable=eval-used
-
-    def _parse_type(self, type_: str) -> type:
-        """Parse the type from the string representation."""
-        type_ = self._make_type_parsable(type_)  # type: ignore
-
-        if get_origin(type_) is Literal:
-            type_ = type(get_args(type_)[0])  # type: ignore
-
-        return type_  # type: ignore
-
-    def _get_nargs(self, type_: type) -> Optional[Union[int, str]]:
-        """Get the nargs for the given type."""
-        if get_origin(type_) is list:
-            return "+"
-        return None
-
-    def _get_choices(self, type_: str, custom_choices: Any) -> Tuple:
-        """Get the choices for the given type."""
-        type_ = self._make_type_parsable(type_)  # type: ignore
-        type_origin = get_origin(type_)
-
-        choices: tuple[Any, ...] = ()
-
-        if type_origin is Literal:
-            choices = get_args(type_)
-
-        if type_origin is list:
-            type_ = get_args(type_)[0]
-
-            if get_origin(type_) is Literal:
-                choices = get_args(type_)
-
-        if type_origin is Union and type(None) in get_args(type_):
-            # remove NoneType from the args
-            args = [arg for arg in get_args(type_) if arg != type(None)]
-            # if there is only one arg left, use it
-            if len(args) > 1:
-                raise ValueError("Union with NoneType should have only one type left")
-            type_ = args[0]
-
-            if get_origin(type_) is Literal:
-                choices = get_args(type_)
-
-        if custom_choices:
-            return tuple(custom_choices)
-
-        return choices
-
-    def build_custom_groups(self):
-        """Build the custom groups from the reference."""
-        for route, v in self.reference.items():
-            for provider, args in v["parameters"].items():
-                if provider == "standard":
-                    continue
-
-                custom_arguments = []
-                for arg in args:
-                    if arg.get("standard"):
-                        continue
-
-                    type_ = self._parse_type(arg["type"])
-
-                    custom_arguments.append(
-                        ArgparseArgumentModel(
-                            name=arg["name"],
-                            type=type_,
-                            dest=arg["name"],
-                            default=arg["default"],
-                            required=not (arg["optional"]),
-                            action="store" if type_ != bool else "store_true",
-                            help=arg["description"],
-                            nargs=self._get_nargs(type_),  # type: ignore
-                            choices=self._get_choices(
-                                arg["type"], custom_choices=arg["choices"]
-                            ),
-                        )
-                    )
-
-                group = ArgparseArgumentGroupModel(
-                    name=provider, arguments=custom_arguments
-                )
-
-                if route not in self.custom_groups:
-                    self.custom_groups[route] = []
-
-                self.custom_groups[route].append(group)
 
 
 class ArgparseTranslator:
